### 2020-04-30 #33 @delsuc
[Rp] Gifa V.4: A complete package for NMR data set processing
**Original article:** [Rp] Gifa V.4: A complete package for NMR data set processing

**Scientific domain:**  NMR, data processing
**Programming language:**  Fortran77, C, Gifa macro
**Suggested editor:**  Pierre de Buyl, Konrad Hinsen, Thomas Arildsen, Georgios Detorakis
