### 2020-05-04 #33 @rougier
@pdebuyl @khinsen Can you serve as editor for this submission for the Ten Years Reproducibility Challenge ?
