### 2020-05-11 #33 @khinsen
@delsuc Could you add instructions for compiling, installing, and running the tests? The code archive looks like this should be straightforward, but it isn't obvious how to start. Also, the paper suggests that running the code requires a 32-bit system. That should also be stated in the README so that people don't waste time trying anything else.
