### 2020-05-11 6a4b4d5
Update README.md with build recipe
