### 2020-05-11 #33 @delsuc
Thanks @khinsen , that was needed, and I should have done it in the first hand.
I added a part in the README which shows how to build the binary, and reproduce the figures in the manuscript.
