### 2020-05-12 #33 @khinsen
Thanks @delsuc, that helps a lot!

Following the instructions, I almost managed to compile Gifa. I had to install the package `libxext-dev` in order to succeed. Then I could run the tests.

I haven't been able so far to reproduce the figures. Here's what I do:
1) `cd Gifa/article`
2) `gifa` (this opens a few windows)
3) `fig2a.g` (this open a dialog box entitled "Display control")
4) Push the OK button.

I get the promised error message about `dispcont_doit`, but no figure. All windows remain blank.

Next:
1) `cd Gifa/article`
2) `gifa` (this opens a few windows)
3) `figure3.g` (this open a dialog box entitled "figure_3", asking me to close the 'gifa' bar and then the dialog)
4) I close the 'gifa' bar (by closing the window)
5) I close the dialog via the OK button.
6) Error message in the terminal: 'No such file or directory. Aborting execution of command file : figure3.g  At line : 11  Error in processing command : READ'

It looks like the macro loads a data file under `Gifa/data`, so I tried again from my home directory (which contains `Gifa`), and then I do get figure 3.

BTW, it would be useful to tell adventurous explorers that "exit" is the command to get out of Gifa!
