### 2020-05-12 921f04f
extended the README
