+++ README.md
@@ -7,7 +7,7 @@ Jean-Luc Pons, Thérèse E. Malliavin, Marc A. Delsuc Gifa V. 4: A complete pack
 *Journal of Biomolecular NMR* (1996), Volume **8**, Issue 4, pp 445–452 
 doi: [10.1007/BF00228146](https://doi.org/10.1007/BF00228146)
  
-**Open Access version:** http://www.delsuc.net/pdf/2007-Journal%20of%20Magnetic%20Resonance-Tramesel.pdf
+**pdf version:** https://link.springer.com/content/pdf/10.1007/BF00228146.pdf
 
 It is also feature an [article](article/article.pdf) submitted as a contribution to the ["Ten Years Reproducibility Challenge"](https://github.com/ReScience/ten-years) run by [**ReScience C**](https://rescience.github.io/).
 
