### 2020-05-12 #33 @delsuc
Thank you @khinsen for pointing out these errors !

- install `libxext-dev` - of course - I just forgot to copy it in the README.

- `fig2a.g` is really badly named ! it actually contains the listing in Figure 2a of the original paper, and reproduces Figure 2c there, which is Figure 1 from the Rescience work (sorry for the strange numbering!) - I considered that there is no need to reproduce Figure 2a and 2b from the original as these are just plain listings.
So the appearance of the dialog box is the whole purpose - I realise this is a bit strange, so I added additional resources so that the correct behaviour is reproduced as well.

- the README has been extended with basic *Gifa* usage - and how to reproduce the figures.
