### 2020-05-13 #33 @khinsen
Thanks @delsuc, now everything works satisfactorily for me. So here comes my review! (pinging @pdebuyl !)

I enjoyed reading this article, which describes the resurrection of a piece of software from the late 1980s/early 1990s that was developed using the top state of the art of that time, paying careful attention to both efficiency and usability. What this reproduction illustrates is (1) the stability of the technology of the time, (2) in particular for software that was from the start written to be portable between the many Unix flavors of the day. The only critical aspect from a long-term reproducibility point of view is the reliance of the memory management architecture on a 32-bit address space.

There is one minor mistake in the paper: there is no "GNU foundation". That should probably be the [GNU Project](https://www.gnu.org/), which is supported by the [Free Software Foundation](https://www.fsf.org/).

The paper is readable as-is but could benefit from a round of careful proofreading for typos etc.
