### 2020-05-13 #33 @delsuc
Thank you @khinsen for this nice review and for helping in improving the quality of the software as well as the manuscript.
 I will correct the GNU point, and will go carefully through the text for proof reading.
