+++ article/content.tex
@@ -1,11 +1,11 @@
 \hypertarget{introduction}{%
 \section{Introduction}\label{introduction}}
 
-I present here the reproduction of the 1996 article ``Gifa V.4: A
+I present here the reproduction of the 1996 article entitled ``Gifa V.4: A
 complete package for NMR data set processing'' \cite{Pons_1996}.
 
 The \emph{Gifa} project is a carrier long effort. It was started in 1987
-as a way to gather various codes and ideas I had developed to handled
+as a way to gather various codes and ideas I had developed to handle
 and process NMR datasets in the early days of 2D and 3D NMR\footnote{I
   started doing NMR as early as 1979, for my French DEA project. 2D NMR
   was in its infancy (the initial paper from R.Ernst group is dated
@@ -18,7 +18,7 @@ and process NMR datasets in the early days of 2D and 3D NMR\footnote{I
 has been built over a careful memory management, a tuned data
 organisation, and a large set of elementary functions meant to be
 pipelined. The version 4 published in 1996 was a major improvement over
-the previous published version (1988) as it brought a complete scripting
+the previously published version (1988) as it brought a complete scripting
 language, a complete graphic user interface based on the Motif library,
 and a database management system. This program was not open-source at
 that time, but nevertheless profited from a large number of users, and
@@ -29,15 +29,15 @@ make it freely available on a code forge.
 \hypertarget{Historical-context}{%
 \section{Historical context}\label{Historical-context}}
 When I entered the \emph{ten years challenge}, I first thought of
-reproducing a latter avatar of the same project, a 2007 program
+reproducing a later avatar of the same project, a 2007 program
 implementation called at that time NPK \cite{Tramesel_2007}. The NPK
 version was an effort to move away from the monolithic environment of
 \emph{Gifa} which was nearly impossible to port to MacOs or Windows.
 This had lead us to move the graphic from X/Motif to Java/Swing and the
-scripting language to python 2.1 (thanks to then excellent
+scripting language to python 2.1 (thanks to the excellent
 \texttt{jython} package recently released then). However the
 mathematical core was unchanged and still written in F\textsc{ortran}77
-with strong 32 bit limitations. Because of this limitation, NPK was
+with strong 32 bit limitations. Because of these limitations, NPK was
 abandoned, and a much extended program has been developed in python
 since then \cite{Spike2016}.
 
@@ -47,8 +47,8 @@ effort, in particular considering my poor knowledge of Java and the fact
 that nobody was using this program any more. This is why I decided to
 fall back on the 1996 version of the same project.
 
-\emph{Gifa} was started in the eighties, when the first 32~bits
-architectures (VMS) were coexisting with older 16~bits ones (PDP-11). It
+\emph{Gifa} was started in the eighties, when the first 32~bit
+architectures (Vax/VMS) were coexisting with older 16~bit ones (PDP-11). It
 was also a time were 64MB of memory was plenty, and the challenge was to
 fit the 2D NMR datasets in memory, in contrast with other programs at
 that time which were doing lengthy processes by swapping the data on
@@ -67,13 +67,13 @@ operations (Fourier transforms, data-set arithmetic, model fitting,
 \ldots{}) and some advanced tools (Linear Prediction tools, Maximum
 Entropy optimisation, DOSY processing by Inverse Laplace
 transform\ldots{}). It is able to process 1D, 2D and in a less optimized
-manner 3D data-sets. It comes with a complete recursive script language
-with variable scoping, recursion, arrays, dictionaries, file I/O, and
+manner 3D data-sets. It comes with a complete scripting language
+with variable scoping, (limited) recursion, arrays, dictionaries, file I/O, and
[…]
