+++ article/metadata.yaml
@@ -10,7 +10,7 @@ title: "[Rp] Gifa V.4: A complete package for NMR data set processing"
 # List of authors with name, orcid number, email and affiliation
 # Affiliation "*" means contact author (required even for single-authored papers)
 authors:
-  - name: Marc-André Delsuc
+  - name: Marc-Andr{\'e} Delsuc
     orcid: 0000-0002-1400-5326
     email: madelsuc@unistra.fr
     affiliations: 1,2,*
@@ -25,7 +25,7 @@ affiliations:
     
   - code:    2
     name:    CASC4DE,  SAS
-    address: Illkirc, France
+    address: Illkirch, France
     
 # List of keywords (adding the programming language might be a good idea)
 keywords: NMR, data processing, FORTRAN77
@@ -83,7 +83,7 @@ language: FORTRAN77
 
 # For example, the URL of the GitHub issue where review actually occured
 review: 
-  - url: 
+  - url: https://github.com/ReScience/submissions/issues/33
 
 contributors:
   - name:
