### 2020-05-18 #33 @delsuc
@pdebuyl I did my best to correct typos and improve the text. It is now available on the repository

I have two remarks:

- for some reason my name and affiliation do not appear in the pdf header, and the compilation always ends-up with an error. I double checked the metadata and the tex, but I could not find where is the problem.
- in which order should I/We register to swh and Zenodo DOI .
