+++ article/content.tex
@@ -177,7 +177,7 @@ available for the distribution 16.04, but the GNU Project proposes \texttt{gdbm}
 adapted to a modern system and had to be adapted.
 
 \begin{itemize}
-\tightlist
+%\tightlist
 \item
   At that time, \texttt{perl} was in version 4 (\texttt{perl} 5 was
   released in 1994 and not widely adopted then). Ubuntu 16.04
