### 2020-05-18 2f49161
Replace accented characters by LaTeX equivalents
