+++ article/metadata.yaml
@@ -20,7 +20,7 @@ authors:
 # contributions" as name (with no address).
 affiliations:
   - code:    1
-    name:    Institut de Génétique Biologie Moléculaire et Cellulaire - UMR 7104 - Inserm U 1258
+    name:    Institut de G{\'e}n{\'e}tique Biologie Mol{\'e}culaire et Cellulaire - UMR 7104 - Inserm U 1258
     address: Illkirch, France
     
   - code:    2
