### 2020-05-18 #33 @khinsen
@delsuc Please see https://github.com/delsuc/Gifa/pull/1 for the small corrections required to compile the article.

As for the archive, you only need to request SWH to archive your repository, no registration required. The request form is at https://archive.softwareheritage.org/save/. Just paste the git reference for your repository into the field at the bottom and push "submit". It can take a few hours before your request is processed. You can check progress on the tab "Browse save requests". You don't need to deposit anything on Zenodo, that's the editors' job.
