### 2020-06-01 #33 @pdebuyl
Hi @delsuc ,

A few minor comments:
1. I don't know the expression "a carrier long effort" (in "Introduction"), I would think that it is not commonly used in English.
2. There is a bit of chance in having a compatibility mode from "gdbm" for "ndbm", maybe this could be stated in the conclusion.
3. Question: is there any code taken from the Numerical Recipes?

Very interesting read! The abandonment of the java version over the Fortran/C one is speaking!

Out of curiosity, is gifa in use in current research projects?

I will proceed to the publication this week. Comment 1 is optional (I am not a native speaker myself), as well as comment 2.
