### 2020-06-17 #33 @delsuc
Hi @pdebuyl ,
Thank you for the comments. I reply here to each points
1.  this is a silly error of mine - the correct spelling is of course "a career-long effort"
2. a sentence has been added in the conclusion (and some minor corrections as well)
3. Numerical Recipes (N.R.) has been used in several places
      - in the Fourier transform codes (Cpx and Real FFT and inverse FFT) - however the logic of the algorithm was modified to adapt causality in Maximum Entropy optimisation
      - in the powell optimiser, but the code was slightly adapted
      - in the random number generator - unmodified it seems
      - then Brent of Levenberg-Marquardt methods, cited in the text are *inspired* from N.R. but rewritten to my "hand"

The text has been modified following points 1 and 2.
