### 2020-06-17 91f842f
minor corrections in article, following @pdebuyl advices
