+++ article/README.md
@@ -1,30 +1,13 @@
-### [ReScience C](https://rescience.github.io/) article template
+### Reproducibility report for "Gifa V.4: A complete package for NMR data set processing"
 
-This repository contains the Latex (optional) template for writing a ReScience
-C article and the (mandatory) YAML metadata file. For the actual article,
-you're free to use any software you like as long as you enforce the proposed
-PDF style. A tool is available for the latex template that produces latex
-definitions from the metadata file. If you use another software, make sure that
-metadata and PDF are always synced.
+**Original article:**
 
-You can also use overleaf with the [provided template](https://www.overleaf.com/read/kfrwdmygjyqw) but in this case, you'll have to enter `metadata.tex` manually.
-
-#### Usage
-
-For a submission, fill in information in
-[metadata.yaml](./metadata.yaml), modify [content.tex](content.tex)
-and type:
-
-```bash
-$ make 
-```
-
-This will produce an `article.pdf` using xelatex and provided font. Note that you must have Python 3 and [PyYAML](https://pyyaml.org/) installed on your computer, in addition to `make`.
+Jean-Luc Pons, Thérèse E. Malliavin, Marc A. Delsuc Gifa V. 4: A complete package for NMR data set processing. 
+*Journal of Biomolecular NMR* (1996), Volume **8**, Issue 4, pp 445–452 
+doi: [10.1007/BF00228146](https://doi.org/10.1007/BF00228146)
+ 
+**pdf version:** https://link.springer.com/content/pdf/10.1007/BF00228146.pdf
 
 
-After acceptance, you'll need to complete [metadata.yaml](./metadata.yaml) with information provided by the editor and type again:
-
-```bash
-$ make
-```
+This article has been submitted to the ["Ten Years Reproducibility Challenge"](https://github.com/ReScience/ten-years) run by [ReScience C](https://rescience.github.io/).
 
