### 2020-06-18 #33 @pdebuyl
Thank you @delsuc for the reply and updates. This is my first paper with the new editorial scripts for ReScience, I'll get to it.
