### 2020-06-23 #33 @pdebuyl
@ReScience/editors / @ReScience/reviewers are people around here able to access zenodo's sandbox? I can't login since three days and would like to proceed with the publication.
