### 2020-06-23 #33 @otizonaizit
The zenodo sandbox was not working for me either last week as I published two ReScience papers. I just crossed my fingers and went directly for the real thing…
