### 2020-06-24 #33 @pdebuyl
Thank you @delsuc for the merges. The file is online on zenodo https://zenodo.org/record/3904595#.XvM5xHVfhhE

@rougier I filed a PR to the website for adding the bib entry.

Thanks @khinsen for the review :-)
